requirements.txt
fastapi
uvicorn
openai
TTS
python-multipart
python-dotenv